Remove root-* pages differentiation. Cleanup files
